Specify the priority for extension-state based rules
Some declarativeNetRequest rules that the DuckDuckGo extension
generates do not relate directly to an extension
configuration (e.g. the block list), but rather to the extension's
state (e.g. the user clicked to allowlist a website). Let's add a
place to put the priorities for those rules, so that the relative
priorities for all declarativeNetRequest rules can be properly
tested.