Add contentBlocking, unprotectedTemporary and denylisting support
The extension configuration (extension-config.json) has several
different ways for privacy protections to be disabled for a
website. Important for when a website is broken by our protections. We
need to support those for Chrome MV3. So far, we support the
trackerAllowlist[1], but not the contentBlocking[2] or
unprotectedTemporary[3] allowlists.

Furthermore, the user is allowed to override the contentBlocking and
unprotectedTemporary allowlists by "denylisting" a website. We need to
add Chrome MV3 support for that too.

1 - https://github.com/duckduckgo/privacy-configuration/blob/main/features/tracker-allowlist.json
2 - https://github.com/duckduckgo/privacy-configuration/blob/main/features/content-blocking.json
3 - https://github.com/duckduckgo/privacy-configuration/blob/main/features/unprotected-temporary.json