Make the tds CLI easier to use
When the CLI was originally designed, the idea was for a
configuration (either block list or extension configuration) to be the
input, and a ruleset to be the output. As we ported more features
however, that was slowly stretched by needing to pass in things like
the list of supported surrogate scripts and the list of user
"denylisted" domains.

Also, while the CLI is often useful during testing/development, it
hasn't been used as much more recently. That meant it has fallen
behind as we've made other changes. While it's still not a priority to
update, let's take a minute to at least make the list of supported
surrogate scripts easier to specify.